Added list and basic core object iterator, sequence and deep GC support
git-svn-id: svn://falconpl.org/falcon/core/trunk@95 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ rtl/math.cpp
@@ -280,7 +280,7 @@ FALCON_FUNC  flc_fract ( ::Falcon::VMachine *vm )
    Item *num = vm->param( 0 );
    if ( num->type() == FLC_ITEM_INT )
    {
-      vm->retval( 0.0 );
+      vm->retval( (int64) 0 );
    }
    else if ( num->type() == FLC_ITEM_NUM )
    {
